created test utils and move some data_utils_test functions in
