+++ ark/utils/test_utils.py
@@ -1,0 +1,29 @@
+import os
+import numpy as np
+import skimage.io as io
+
+
+# required metadata for mibitiff writing (barf)
+METADATA = {
+    'run': '20180703_1234_test', 'date': '2017-09-16T15:26:00',
+    'coordinates': (12345, -67890), 'size': 500., 'slide': '857',
+    'fov_id': 'Point1', 'fov_name': 'R1C3_Tonsil',
+    'folder': 'Point1/RowNumber0/Depth_Profile0',
+    'dwell': 4, 'scans': '0,5', 'aperture': 'B',
+    'instrument': 'MIBIscope1', 'tissue': 'Tonsil',
+    'panel': '20170916_1x', 'mass_offset': 0.1, 'mass_gain': 0.2,
+    'time_resolution': 0.5, 'miscalibrated': False, 'check_reg': False,
+    'filename': '20180703_1234_test', 'description': 'test image',
+    'version': 'alpha',
+}
+
+
+def _create_img_dir(temp_dir, fovs, imgs, img_sub_folder="TIFs", dtype="int8"):
+    tif = np.random.randint(0, 100, 1024 ** 2).reshape((1024, 1024)).astype(dtype)
+
+    for fov in fovs:
+        fov_path = os.path.join(temp_dir, fov, img_sub_folder)
+        if not os.path.exists(fov_path):
+            os.makedirs(fov_path)
+        for img in imgs:
+            io.imsave(os.path.join(fov_path, img), tif)
